fix(packets): make cinfo icon short

+++ Packets/NosSmooth.Packets/Server/Players/CInfoPacket.cs
@@ -61,7 +61,7 @@ public record CInfoPacket
     [PacketIndex(10)]
     PlayerClass Class,
     [PacketIndex(11)]
-    byte Icon,
+    short Icon,
     [PacketIndex(12)]
     short Compliment,
     [PacketIndex(13)]
